Update readme file

+++ src/Doctrine/Rest/Action/DeleteAction.php
@@ -14,7 +14,7 @@ class DeleteAction extends RestAction
     public function handle(RestRequest $request)
     {
         $entity = $this->repository()->findByIdentifier($request);
-        $request->authorize($entity);
+        $this->authorize($request, $entity);
 
         $this->repository()->em()->remove($entity);
         $this->repository()->em()->flush();
